refactor: rename useUsersData file to useCreateUser
